Include all code from the mixin directly in secure-payment-form
